requirements.txt
# To ensure app dependencies are ported from your virtual environment/host machine into your container, run 'pip freeze > requirements.txt' in the terminal to overwrite this file
requests==2.27.1
discord==2.1.0
PyYAML==6.0
tabulate==0.9.0